산재GPT-API/requirements.txt
fastapi
uvicorn[standard]
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
faiss-cpu
pypdf
python-dotenv
pydantic
